separate out AuthProvider tests
